Icon menu basic functionality
