+++ HouseholdManager/Views/Room/Create.cshtml
@@ -19,7 +19,8 @@
             </div>
             <div class="form-group">
                 <label asp-for="Icon" class="control-label"></label>
-                <input asp-for="Icon" class="form-control" />
+                <input asp-for="Icon" id="iconInput" class="form-control" hidden />
+                @await Html.PartialAsync("_IconMenuPartial");
                 <span asp-validation-for="Icon" class="text-danger"></span>
             </div>
             <div class="form-group">
@@ -28,6 +29,7 @@
         </form>
     </div>
 </div>
+
 
 <div>
     <a asp-action="Index">Back to List</a>
